pyproject.toml
[tool.poetry]
name = "ASTRA"
version = "0.1.0"
description = "Generation of semantically enriched mobility data from travel surveys."
authors = ["Matthias Mohn", "Maja Schneider"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
ipykernel = "^6.24.0"
sentence-transformers = "^2.3.1"
osmnx = "^1.9.4"
rasterio = "^1.3.9"
pyarrow = "^15.0.2"
ray = {extras = ["default"], version = "^2.10.0"}
modin = "0.26.0"
grpcio = "^1.62.1"
cuda-python = "^12.4.0"
lightning = "^2.2.1"
jupyterlab = "^4.1.5"
orjson = "^3.10.0"
faiss-cpu = "^1.8.0"
xxhash = "^3.4.1"
memory-profiler = "^0.61.0"
wheel = "^0.43.0"
seaborn = "^0.13.2"
pyproj = "^3.6.1"
gluonts = "^0.15.1"
black = "^24.4.2"
notebook = "^7.2.1"
ipywidgets = "^8.1.3"
mlflow = "^2.14.1"
torchinfo = "^1.8.0"
tomli = "^2.0.1"
fiona = "1.9.6"
click = "^8.1.7"
contextily = "^1.6.2"
psycopg = "^3.2.4"
lxml = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
